requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.3
